Lucas/fix prem days and outfit (#16)
* Fix outfit mapping to proto

Outfits were wrongly mapped to proto, so its colors were misplaced. This will properly map the proto properties to the correct database columns.

* Fix uint overflow in premium_unti parameter

PremUntil contains a timestamp of the day when the premium is over. That's a number bigger than uint_32 limit, thus it was overflowing and returning wrong value. This commit updates the protobuf, the structures and methods to uint64.