Event calendar correction
The file you shared is almost correct, but the main problem is with the date format in the XML and how it is interpreted by Go. Your XML is using the format MM/DD/YYYY, but Go expects DD/MM/YYYY or another format depending on the time.Parse function. Here is how you can tweak the code: